Extract reusable EditorColorLegendSection
